Add code generation for event handlers
We want the MumbleClient class (main entry point of the library) to have
easy to access events which handle each type of message. Code generation
is the easiest way to have a maintainable list of these event handlers.

Firing these events is another story. Since the connection class only
gives us the non-generic IMessage, we need a way to map that interface
to a concrete type (GetType()) and then downcast it. Using more code
generation to make a lookup table which maps types to event firing
delegates makes this simple and efficient.