Added background tasks for constant connection.
Making the connection class take cancellationTokens everywhere allows us
to easily cancel the async operations when we need to (on disconnect
mostly). With a means to cancel these operations we can make long
running tasks which handle constantly reading from the network for new
messages and sending pings at regular intervals.

Also, we can add a disconnect method as long as our connection object
can be reset. That's where the connectionFactory comes in.