Clean up the dependencies for client collections
The user object needs to know about the list of channels and the channel
object needs to know about the list of users. The way I the base model
hold a reference to its containing collection didn't make sense so now
they depend on the client itself. Much easier relationship