Add ability to capture message responses
Some mumble message are expected to have a corresponding server
response. As an example, a user state message which moves a user to a
channel is responded to by either an identical user state message
showing the user has moved or a permission denied message. This pattern
is somewhat common so functionality was added to the client to support
it.

Because of the existing design, this was accomplished by creating a
method which can wait for an event fire (WaitForMessageAsync<T>)